fixes to refactoring
