+++ pages/mentors.page.js
@@ -3,10 +3,10 @@ import axios from "axios";
 import dynamic from "next/dynamic";
 import MentorCard from "../components/mentor";
 import teamsData from "./data/teamsData";
-import { useApi } from "./customHook/useAPi.js";
+import { useApi } from "../hook/useAPi.js";
 import { list } from "postcss";
-const Header = dynamic(() => import("../components/Header"));
-const SimpleBanner = dynamic(() => import("../components/SimpleBanner"));
+const Header = dynamic(() => import("../components/layout/Header"));
+const SimpleBanner = dynamic(() => import("../components/basic/SimpleBanner"));
 
 function Mentors({ mentorsData }) {
   const [query, setQuery] = useState("");
