+++ pages/privacy.page.js
@@ -1,7 +1,7 @@
 import React from "react";
 import dynamic from "next/dynamic";
-import Header from "../components/Header";
-const Footer = dynamic(() => import("../components/Footer"));
+import Header from "../components/layout/Header";
+const Footer = dynamic(() => import("../components/layout/Footer"));
 import Head from "next/head";
 
 function privacy() {
