+++ pages/resetpassword/[entity]/[resetToken].page.js
@@ -1,9 +1,9 @@
 import React from "react";
 import dynamic from "next/dynamic";
-const Header = dynamic(() => import("../../../components/Header"));
+const Header = dynamic(() => import("../../../components/layout/Header"));
 import ResetPassword from "../../../components/resetpassword";
 
-const Footer = dynamic(() => import("../../../components/Footer"));
+const Footer = dynamic(() => import("../../../components/layout/Footer"));
 
 function Index({ resetToken }) {
   return (
