+++ pages/userAuth.page.js
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
 import { useRouter } from "next/router";
-import Header from "../components/Header";
+import Header from "../components/layout/Header";
 import Login from "../components/user/login";
 import Register from "../components/user/register";
 import Head from "next/head";
